implementacion comprar pasajes

+++ src/components/cart/index.js
@@ -2,7 +2,7 @@ import React, { Fragment, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { cartLoad } from '../../redux/actions/cartAction';
 import './Cart.css'
-import { deleteAll, cartLoadError, deleteFlightReservation } from '../../redux/actions/cartAction'
+import { deleteAll, cartLoadError, deleteFlightReservation, buyTicket } from '../../redux/actions/cartAction'
 
 import Button from '@material-ui/core/Button';
 import Table from '@material-ui/core/Table';
@@ -11,7 +11,8 @@ import TableHead from '@material-ui/core/TableHead';
 import TableContainer from '@material-ui/core/TableContainer';
 import Paper from '@material-ui/core/Paper';
 import TableBody from '@material-ui/core/TableBody';
-import { useStyles, StyledTableCell, StyledTableRow } from './style'
+import { useStyles, StyledTableCell } from './style'
+import Swal from 'sweetalert2'
 
 export default function Cart() {
   const dispatch = useDispatch();
@@ -19,20 +20,23 @@ export default function Cart() {
   const flights = useSelector(state => state.cartReducer.flights);
   const error = useSelector(state => state.cartReducer.error)
   const loading = useSelector(state => state.cartReducer.loading)
-  
+
+  const buyTickets = (ticket) => dispatch(buyTicket(ticket))
+
   useEffect(() => {
-    dispatch(cartLoad());
-    if (flights.length === 0) {
-      console.log('golaaaaaaa')
+    dispatch(cartLoad())
+    if (flights === []) {
       dispatch(cartLoadError())
     }
+    console.log(flights)
 
-    
+
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const sumaTotal = () => {
     let total = 0
+
     Array.from(flights).forEach(flight =>
       total += flight.price)
 
@@ -40,19 +44,57 @@ export default function Cart() {
   }
 
   const deleteFlight = id => {
-    dispatch(deleteFlightReservation(id))
-  }
 
+    Swal.fire({
+      title: 'Estas Seguro?',
+      text: "un pasaje que se elimina no se puede recuperar",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'si, eliminar!',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.value) {
+        dispatch(deleteFlightReservation(id))
+      }
+    })
+
+  }
 
   const deleteAllflights = () => {
     dispatch(deleteAll())
   }
+
+  const buyTicketsFlights = e => {
+    e.preventDefault()
+
+    Swal.fire({
+      title: 'Estas Seguro?',
+      text: "no se puede cancelar la compra",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, comprar!',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.value) {
+        flights.forEach(flight =>
+          buyTickets(
+            flight
+          )
+        )
+      }
+    })
+
+  }
   return (
     <Fragment>
 
-      {error ? <p>hubo un error</p>:null   }
+      {error ? <p>hubo un error</p> : null}
 
-      {loading ? <p>hubo un error</p>:null   }
+      {loading ? <p>hubo un error</p> : null}
       <TableContainer component={Paper}>
         <Table className={classes.table} aria-label="customized table">
           <TableHead>
@@ -61,19 +103,18 @@ export default function Cart() {
               <StyledTableCell align="center">Destino</StyledTableCell>
               <StyledTableCell align="center">Salida</StyledTableCell>
               <StyledTableCell align="center">Aerolinea</StyledTableCell>
-              <StyledTableCell align="center" >Asiento</StyledTableCell>
-              <StyledTableCell align="center" >Clase</StyledTableCell>
-              <StyledTableCell align="center" >Precio</StyledTableCell>
-              <StyledTableCell align="center" >Accion</StyledTableCell>
+              <StyledTableCell align="center">Asiento</StyledTableCell>
+              <StyledTableCell align="center">Clase</StyledTableCell>
+              <StyledTableCell align="center">Precio</StyledTableCell>
+              <StyledTableCell align="center">Accion</StyledTableCell>
             </TableRow>
           </TableHead>
           <TableBody>
-            
-            {Array.from(flights).length=== 0 ? 'no hay pasajes' : (Array.from(flights).map(flight =>
-              <StyledTableRow key={flight.id}>
 
+              {Array.from(flights).length === 0 ?'no hay pasajes':(Array.from(flights).map(flight =>
+              <TableRow key={flight.id}>
                 <StyledTableCell align="center">{flight.origin}</StyledTableCell>
-                <StyledTableCell align="center" >{flight.destination}</StyledTableCell>
+                <StyledTableCell align="center">{flight.destination}</StyledTableCell>
                 <StyledTableCell align="center">{flight.exit}</StyledTableCell>
                 <StyledTableCell align="center">{flight.airport}</StyledTableCell>
                 <StyledTableCell align="center">{flight.seat}</StyledTableCell>
@@ -86,13 +127,14 @@ export default function Cart() {
                     onClick={() => deleteFlight(flight.id)}
                   >Eliminar</Button>
                 </StyledTableCell>
-              </StyledTableRow>
+              </TableRow>
 
             ))}
           </TableBody>
         </Table>
 
       </TableContainer>
+      <form onSubmit={buyTicketsFlights}>
 
         <div className="botones">
           <Button variant="contained" color="primary" onClick={() => deleteAllflights()}>Limpar carro</Button>
@@ -102,8 +144,9 @@ export default function Cart() {
           <div className="botonVolver">
             <Button variant="contained" color="secondary">Volver</Button>
           </div>
-          <Button variant="contained" color="primary">Comprar</Button>
+          <Button type="submit" variant="contained" color="primary" >Comprar</Button>
         </div>
+      </form>
 
     </Fragment>
   )
